pyproject.toml
[project]
name = "values"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pillow>=12.1.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-spec>=6.0.0",
    "requests>=2.33.0",
]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=term-missing --spec"
